mapper/requirements.txt
Django==2.0
djangorestframework==3.7.3
pkg-resources==0.0.0
pygeoif==0.7
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
